Refactoring code a bit, adding tests too. Right now they don't get stuck though, so need to examine closer on actual situation when the pipeline got stuck.
