+++ pipelines/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@data-heaving/orchestration",
-  "version": "0.9.0",
+  "version": "0.9.1",
   "author": {
     "name": "Stanislav Muhametsin",
     "email": "346799+stazz@users.noreply.github.com",
@@ -20,7 +20,7 @@
     "run": "node dist/cjs",
     "dev-run": "ts-node dist/cjs",
     "test": "npm run run-test && npm run remove-test-files",
-    "run-test": "echo TODO TESTS",
+    "run-test": "ava",
     "remove-test-files": "rm -rf dist/*/tests src/*/tests"
   },
   "dependencies": {
@@ -30,6 +30,7 @@
     "@types/node": "14.14.31",
     "@typescript-eslint/eslint-plugin": "4.22.0",
     "@typescript-eslint/parser": "4.22.0",
+    "ava": "4.0.0-alpha.2",
     "eslint": "7.25.0",
     "eslint-config-prettier": "8.3.0",
     "eslint-plugin-prettier": "3.4.0",
@@ -42,5 +43,11 @@
     "src/"
   ],
   "main": "./dist/cjs/index.js",
-  "module": "./dist/esm/index.js"
+  "module": "./dist/esm/index.js",
+  "ava": {
+    "files": [
+      "dist/cjs/**/tests/*.js"
+    ],
+    "timeout": "10m"
+  }
 }
